requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt
pwdlib[argon2]
python-multipart
pydantic[email]
